requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.7.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-jose[cryptography]>=3.3.0
gunicorn>=21.2.0
whitenoise>=6.6.0
drf-yasg>=1.21.7
python-dotenv>=1.0.0
requests-cache==1.1.1 